Moved the logic to enumerate the savegames to SaveGameProvider (away from GameFileArray)
